fix: use express router instead of application
Previously, this package created an express application, and attached the /robots.txt route to that
application. Generally, this isn't a useful feature for real application, and it can cause
x-powered-by to be visible on (all) response requests. Instead of using an application, switch the
package to use a router.

BREAKING CHANGE: The main export now returns an express router instead of an express application.